% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_utils.R
\name{build_namespace}
\alias{build_namespace}
\title{Build project namespace}
\usage{
build_namespace(project)
}
\arguments{
\item{project}{Character. The project code.}
}
\description{
Most project files are saved under \verb{/data/repository/\{project\}}.
Some networks, namely ATAP, MigraMar and NEP, are directly hosted by OTN and
so have their files saved under \verb{/data/repository/\{network\}/\{project\}}. This
function does some light cleaning of the project code and adjusts for these
OTN-hosted networks.
}
\keyword{internal}
\keyword{utils}
