% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_checks.R
\name{check_server}
\alias{check_server}
\title{Check to see if the user has provided a server}
\usage{
check_server(server)
}
\arguments{
\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}
}
\description{
Check to see if the user has provided a server
}
\keyword{internal}
