% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otn_api.R
\name{.otn_api}
\alias{.otn_api}
\title{Call OTN API endpoints.}
\usage{
.otn_api(endpoint, server = otn_global$server_url)
}
\arguments{
\item{endpoint}{Character. API endpoint to call.}

\item{server}{URL of the network's Plone CMS}
}
\description{
Call OTN API endpoints.
}
\keyword{internal}
