% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{.otn_search}
\alias{.otn_search}
\title{Internal function for searching OTN projects.}
\usage{
.otn_search(server, type, search_term)
}
\arguments{
\item{server}{Character. Lowercase network code of the desired server. One of
"otn", "act", "npact", or "devel" (the OTN development server). Note that
"etn", "fact", and "glatos" are accepted, but only to produce an error and redirect you.}

\item{type}{Character. Internal data base to search.}

\item{search_term}{Character. Search term to use.}
}
\description{
Internal function for searching OTN projects.
}
\keyword{internal}
