% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_utils.R
\name{is_logged_in}
\alias{is_logged_in}
\title{Check if user is logged in}
\usage{
is_logged_in()
}
\description{
Checks for the presence of a session token in the session environment, erroring
if one does not exist.
}
\keyword{internal}
\keyword{utils}
