% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_extract_files.R
\name{list_extract_files}
\alias{list_extract_files}
\title{List OTN data extraction files}
\usage{
list_extract_files(
  project,
  detection_type = c("all", "matched", "external", "qualified", "sentinel",
    "unqualified"),
  since = NULL
)
}
\arguments{
\item{project}{The project code.}

\item{detection_type}{one of, or a vector of, "all" (default), "matched",
"external", "qualified", "sentinel_tag", or "unqualified". Partial
matching is allowed, and will repair to the correct argument if spaces
or the words detection(s)" are included.
More information on data types can be found on
\href{https://members.oceantrack.org/data/otn-detection-extract-documentation-matched-to-animals}{OTN's website}.}

\item{since}{Currently unused. Future documentation: Only list files
uploaded after this date. Optional, but must be in YYYY-MM-DD format.}
}
\value{
A data frame with columns of "project", "detection_type",
'detection_year', 'file_name', and "url".
}
\description{
List OTN data extraction files
}
