% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_project_files.R
\name{list_project_files}
\alias{list_project_files}
\title{List OTN project files}
\usage{
list_project_files(project)
}
\arguments{
\item{project}{The project code.}
}
\description{
This function lists the file names, types, upload date, and URLs of OTN
project files -- basically everything you see in the \emph{Data and Metadata}
section of your project page. You will be prompted to log in if it is not a
public project.
}
