% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{otn_download}
\alias{otn_download}
\title{Download file(s) from an OTN-style Plone server.}
\usage{
otn_download(files = NULL, url = NULL, outdir = ".")
}
\arguments{
\item{files}{a data.frame returned by by \code{otn_project_files} or
\code{otn_extract_files}. Can have been filtered or edited, but must have columns
of "url" and "name".}

\item{url}{The URL of the file as returned by \code{otn_project_files} or
\code{otn_extract_files}}

\item{outdir}{The output directory (where you want the file to be saved).
Defaults to the current working directory.}
}
\value{
Silently returns a vector of file locations.

#' @seealso
\itemize{
\item Plone REST API documentation:
\itemize{
\item \href{https://6.docs.plone.org/plone.restapi/docs/source/usage/serialization.html#download-serialization}{Serialization}
}
}
}
\description{
Either \code{files} or \code{url} must be provided. If \code{files} is provided, files in
the data.frame are downloaded in parallel.
}
