% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{otn_get_file}
\alias{otn_get_file}
\title{Download a file using the file's URL}
\usage{
otn_get_file(file_url, outdir = ".")
}
\arguments{
\item{file_url}{The URL of the file as returned by \code{otn_project_files} or
\code{otn_extract_files}}

\item{outdir}{The output directory (where you want the file to be saved).
Defaults to the current working directory.}
}
\description{
Download a file using the file's URL
}
