% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_general.R
\name{otn_list_institutions}
\alias{otn_list_institutions}
\title{Retrieve OTN institution list.}
\usage{
otn_list_institutions(server = "otn")
}
\arguments{
\item{server}{Character. Lowercase network code of the desired server. One of
"otn", "act", "npact", or "devel" (the OTN development server). Note that
"etn", "fact", and "glatos" are accepted, but only to produce an error and redirect you.}
}
\description{
Retrieve OTN institution list.
}
