% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_general.R
\name{otn_list_projects}
\alias{otn_list_projects}
\title{Retrieve OTN project list.}
\usage{
otn_list_projects(network = "otn", what = c("all", "mine"))
}
\arguments{
\item{network}{Character. Lowercase code of the desired telemetry network. One of
"otn", "act", "atap", "itag", "migramar", "nep", "npact", "pirat", or
"otn_devel" (the OTN development server). Note that "etn", "fact", "glatos",
"path" and "raft" are accepted, but only to produce an error and redirect you.}

\item{what}{Which projects do you wish to see: \code{all} or \code{mine}? \code{all} can be
viewed without logging in, but you must log in before viewing yours (well, \code{mine}).}
}
\description{
Retrieve OTN project list.
}
