% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{otn_list_species}
\alias{otn_list_species}
\title{Retrieve OTN species list}
\usage{
otn_list_species(server = NULL)
}
\arguments{
\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}
}
\description{
Retrieve OTN species list
}
