% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_files.R
\name{otn_project_files}
\alias{otn_project_files}
\title{List OTN project files.}
\usage{
otn_project_files(project, text = NULL, since = NULL, batch_size = 25)
}
\arguments{
\item{project}{Character. The project code.}

\item{text}{Character. Text to search for in the file name/title. See Details
for more information.}

\item{since}{Character. Filter for files modified since this date
(YYYY-MM-DD format).}

\item{batch_size}{Numeric. The number of results to return. Defaults to 25.}
}
\description{
This function lists the file names, types, upload date, and URLs of OTN
project files -- basically everything you see in the \emph{Data and Metadata}
section of your project page. You will be prompted to log in if it is not a
public project.
}
\seealso{
\link{.otn_files}
}
