% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receiver_summary.R
\name{otn_receiver_summary}
\alias{otn_receiver_summary}
\title{Create summary reports of receiver project data from the OTN data push.}
\usage{
otn_receiver_summary(
  project = NULL,
  server = NULL,
  qualified = NULL,
  unqualified = NULL,
  deployment = NULL,
  ...
)
}
\arguments{
\item{project}{Character. The project code.}

\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}

\item{qualified, unqualified}{Default is NULL: OTN qualified or unqualified
detections in parquet format will be downloaded from the data server. This
argument also accepts a character vector of file paths of your
qualified/unqualified detections if you do not wish to download your files
or your node does not use an OTN-style Plone CMS.}

\item{deployment}{File path of user-supplied master OTN receiver deployment metadata.}

\item{...}{Arguments passed to \code{otndo::make_receiver_push_summary}}
}
\description{
Create summary reports of receiver project data from the OTN data push.
}
\section{No files provided}{


If you only provide your project code and leave all of the arguments as their
defaults, this function will ask you to log in then proceed to download all
of the necessary files. You can speed up this process substantially by
providing already-downloaded files.
}

\section{Output}{


This function creates an HTML report that can be viewed in your web browser.
}

\examples{
\dontrun{
# Using only the project code:
otn_receiver_summary("tailwinds")

# Providing a local file:
otn_receiver_summary("tailwinds", deployment = "my_master_deployment_metadata.xlsx")

# Get a summary fo what has changed since a particular date:
otn_receiver_summary("tailwinds", since = "2022-05-01")
}
}
