% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{otn_search_code}
\alias{otn_search_code}
\title{Search OTN database by project code.}
\usage{
otn_search_code(code, network = "otn")
}
\arguments{
\item{code}{Character. Project code for which to search. Supports partial matching.}

\item{network}{Character. Lowercase code of the desired telemetry network. One of
"otn", "act", "atap", "itag", "migramar", "nep", "npact", "pirat", or
"otn_devel" (the OTN development server). Note that "etn", "fact", "glatos",
"path" and "raft" are accepted, but only to produce an error and redirect you.}
}
\description{
Search OTN database by project code.
}
