% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{otn_search_contact}
\alias{otn_search_contact}
\title{Search OTN database by project contact.}
\usage{
otn_search_contact(contact, network = "otn")
}
\arguments{
\item{contact}{Character. Project contact for which to search. DOES NOT support partial matching.}
}
\description{
Search OTN database by project contact.
}
