% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{otn_search_country}
\alias{otn_search_country}
\title{Search OTN database by project country.}
\usage{
otn_search_country(country, server = "otn")
}
\arguments{
\item{country}{Character. Project country for which to search. DOES NOT support partial matching.}

\item{server}{Character. Lowercase network code of the desired server. One of
"otn", "act", "npact", or "devel" (the OTN development server). Note that
"etn", "fact", and "glatos" are accepted, but only to produce an error and redirect you.}
}
\description{
Search OTN database by project country.
}
