% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{otn_search_species}
\alias{otn_search_species}
\title{Search OTN database by project species}
\usage{
otn_search_species(species, network = "otn")
}
\arguments{
\item{species}{Character. Project species for which to search. DOES NOT
support partial matching and only recognizes scientific names of the
form "Genus species" (note capitalization).}

\item{network}{Character. Lowercase code of the desired telemetry network. One of
"otn", "act", "atap", "itag", "migramar", "nep", "npact", "pirat", or
"otn_devel" (the OTN development server). Note that "etn", "fact", "glatos",
"path" and "raft" are accepted, but only to produce an error and redirect you.}
}
\description{
Search OTN database by project species
}
