% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{otn_set_credentials}
\alias{otn_set_credentials}
\title{Installs your OTN username and password in your \code{.Renviron} file for repeated use.}
\usage{
otn_set_credentials(network, temporary = FALSE, overwrite = FALSE)
}
\arguments{
\item{network}{Character. Lowercase code of the desired telemetry network. One of
"otn", "act", "atap", "itag", "migramar", "nep", "npact", "pirat", or
"otn_devel" (the OTN development server). Note that "etn", "fact", "glatos",
"path" and "raft" are accepted, but only to produce an error and redirect you.}

\item{temporary}{Logical. Scrub credentials after the current session?}

\item{overwrite}{Logical. Overwrite previously-stored credentials?}
}
\description{
This code was adapted from \href{https://github.com/walkerke/tidycensus/blob/ddb33b5f72734a4ff14332bd55cbac4850688600/R/helpers.R}{\code{tidycensus::census_api_key}}. Note that this saves your credentials in your .Renviron, meaning that anyone who is using your computer can theoretically access what your MATOS username and password are. So... use this carefully!
}
\examples{
\dontrun{
otn_set_credentials("act")
}
# Yup, that's it!
}
