% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_summary.R
\name{otn_tag_summary}
\alias{otn_tag_summary}
\title{Create summary reports of receiver project data from the OTN data push}
\usage{
otn_tag_summary(project = NULL, server = NULL, matched = NULL, ...)
}
\arguments{
\item{project}{Character. The project code.}

\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}

\item{matched}{Default is NULL: OTN matched detections in parquet format will
be downloaded from the data server. If you do not wish to download your files
or your node does not use an OTN-style Plone CMS, this argument also accepts
a character vector of file paths of your matched detections. These can be
parquet, CSVs, or zipped folders.}

\item{...}{Arguments passed to \code{otndo::make_tag_push_summary}}

\item{matos_project}{MATOS project number or name that you wish to have summarized}
}
\description{
Create summary reports of receiver project data from the OTN data push
}
\examples{
\dontrun{
# You can just use your ACT project number
otn_tag_summary("mdwea")

# Or provide an optional date to summarize "What's New".
otn_tag_summary("mdwea", since = "2025-11-01")
}
}
