% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{otn_upload}
\alias{otn_upload}
\title{Upload a file to the Plone CMS}
\usage{
otn_upload(file_path, project)
}
\arguments{
\item{file_path}{Character. Location of the file to be uploaded.}

\item{project}{Character. The project code.}
}
\value{
Silently returns a vector of file locations.
}
\description{
Upload a file to the Plone CMS
}
\seealso{
\itemize{
\item Plone REST API documentation:
\itemize{
\item \href{https://6.docs.plone.org/plone.restapi/docs/source/usage/content.html}{Content manipulation}
\item \href{https://6.docs.plone.org/plone.restapi/docs/source/usage/serialization.html#upload-deserialization}{Deserialization}
}
}
}
