% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otn_api.R
\name{.otn_api}
\alias{.otn_api}
\title{Call OTN API endpoints.}
\usage{
.otn_api(endpoint, server = otn_global$server)
}
\arguments{
\item{endpoint}{Character. API endpoint to call.}

\item{server}{Character. Lowercase network code of the desired server. One of
"otn", "act", "npact", or "devel" (the OTN development server). Note that
"etn", "fact", and "glatos" are accepted, but only to produce an error and redirect you.}
}
\description{
Call OTN API endpoints.
}
\keyword{internal}
