% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otn_api.R
\name{.otn_api}
\alias{.otn_api}
\title{Call OTN API endpoints}
\usage{
.otn_api(server = NULL, endpoint)
}
\arguments{
\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}

\item{endpoint}{Character. API endpoint to call.}
}
\description{
Call OTN API endpoints
}
