% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_files.R
\name{.otn_files}
\alias{.otn_files}
\title{Return files associated with an OTN project.}
\usage{
.otn_files(project, since = NULL, batch_size = NULL, type)
}
\arguments{
\item{project}{Character. The project code.}

\item{since}{Character. Filter for files modified since this date
(YYYY-MM-DD format).}

\item{batch_size}{Numeric. The number of results to return. Defaults to 25.}

\item{type}{Character. Portion of the URL representing the data type you wish
to return.}
}
\description{
Return files associated with an OTN project.
}
\seealso{
\itemize{
\item Plone REST API documentation:
\itemize{
\item \href{https://6.docs.plone.org/plone.restapi/docs/source/endpoints/querystringsearch.html}{Querystring Search},
triggered when using the "\code{since}" argument.
\href{https://6.docs.plone.org/plone.restapi/docs/source/endpoints/querystring.html}{Query operations are listed here.}.
\item \href{https://6.docs.plone.org/plone.restapi/docs/source/endpoints/searching.html}{Search},
a simpler search used when the "\code{since}" argument is NULL.
}
}
}
\keyword{internal}
