% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_files.R
\name{.otn_files}
\alias{.otn_files}
\title{Return files associated with an OTN project}
\usage{
.otn_files(project, server = NULL, since = NULL, batch_size = NULL, type)
}
\arguments{
\item{project}{Character. The project code.}

\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}

\item{since}{Character. Filter for files modified since this date
(YYYY-MM-DD format).}

\item{batch_size}{Numeric. The number of results to return. Defaults to 25.}

\item{type}{Character. Portion of the URL representing the data type you wish
to return.}
}
\description{
Return files associated with an OTN project
}
