% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_general.R
\name{.otn_list}
\alias{.otn_list}
\title{Internal function to list small OTN databases.}
\usage{
.otn_list(network, type)
}
\arguments{
\item{network}{Character. Lowercase code of the desired telemetry network. One of
"otn", "act", "atap", "itag", "migramar", "nep", "npact", "pirat", or
"otn_devel" (the OTN development server). Note that "etn", "fact", "glatos",
"path" and "raft" are accepted, but only to produce an error and redirect you.}

\item{type}{Character. Internal data base to retrieve.}
}
\description{
Provides access to the data that is on the main OTN members page. Login is
not required.
}
\keyword{internal}
