% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_general.R
\name{.otn_list}
\alias{.otn_list}
\title{Internal function to list small OTN databases.}
\usage{
.otn_list(server, type)
}
\arguments{
\item{server}{Character. Lowercase network code of the desired server. One of
"otn", "act", "npact", or "devel" (the OTN development server). Note that
"etn", "fact", and "glatos" are accepted, but only to produce an error and redirect you.}

\item{type}{Character. Internal data base to retrieve.}
}
\description{
Provides access to the data that is on the main OTN members page. Login is
not required.
}
\keyword{internal}
