% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{.otn_list}
\alias{.otn_list}
\title{Internal function to list small OTN data bases}
\usage{
.otn_list(server = NULL, type)
}
\arguments{
\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}

\item{type}{Character. Internal data base to retrieve.}
}
\description{
Internal function to list small OTN data bases
}
