% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{.otn_search}
\alias{.otn_search}
\title{Internal function for searching OTN projects.}
\usage{
.otn_search(network, type, search_term)
}
\arguments{
\item{network}{Character. Lowercase code of the desired telemetry network. One of
"otn", "act", "atap", "itag", "migramar", "nep", "npact", "pirat", or
"otn_devel" (the OTN development server). Note that "etn", "fact", "glatos",
"path" and "raft" are accepted, but only to produce an error and redirect you.}

\item{type}{Character. Internal data base to search.}

\item{search_term}{Character. Search term to use.}
}
\description{
Internal function for searching OTN projects.
}
\keyword{internal}
