% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{.otn_search}
\alias{.otn_search}
\title{Internal function for searching OTN projects.}
\usage{
.otn_search(network, type, search_term)
}
\arguments{
\item{type}{Character. Internal data base to search.}

\item{search_term}{Character. Search term to use.}
}
\description{
Internal function for searching OTN projects.
}
\keyword{internal}
