% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{.otn_search}
\alias{.otn_search}
\title{Internal function for searching OTN projects}
\usage{
.otn_search(server, type, search_term)
}
\arguments{
\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}

\item{type}{Character. Internal data base to search.}

\item{search_term}{Character. Search term to use.}
}
\description{
Internal function for searching OTN projects
}
