% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_url.R
\name{.otn_server_url}
\alias{.otn_server_url}
\title{Set the OTN Plone instance with which you wish to interact.}
\usage{
.otn_server_url(
  network = c("otn", "act", "atap", "itag", "migramar", "nep", "npact", "pirat",
    "otn_devel", "etn", "fact", "glatos", "path", "raft"),
  set = TRUE
)
}
\arguments{
\item{network}{Character. Lowercase code of the desired telemetry network. One of
"otn", "act", "atap", "itag", "migramar", "nep", "npact", "pirat", or
"otn_devel" (the OTN development server). Note that "etn", "fact", "glatos",
"path" and "raft" are accepted, but only to produce an error and redirect you.}

\item{set}{Set the network in the current session?}
}
\description{
Set the OTN Plone instance with which you wish to interact.
}
