% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_files.R
\name{otn_extract_files}
\alias{otn_extract_files}
\title{List OTN extract files}
\usage{
otn_extract_files(project, batch_size = NULL, server = NULL)
}
\arguments{
\item{project}{Character. The project code.}

\item{batch_size}{Numeric. The number of results to return. Defaults to 25.}

\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}
}
\description{
This function recursively lists files in the \emph{Detection Extracts} section of
your project.
}
