% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_general.R
\name{otn_list_countries}
\alias{otn_list_countries}
\title{Retrieve OTN country list.}
\usage{
otn_list_countries(network = "otn")
}
\arguments{
\item{network}{Character. Lowercase code of the desired telemetry network. One of
"otn", "act", "atap", "itag", "migramar", "nep", "npact", "pirat", or
"otn_devel" (the OTN development server). Note that "etn", "fact", "glatos",
"path" and "raft" are accepted, but only to produce an error and redirect you.}
}
\description{
Retrieve OTN country list.
}
