% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_general.R
\name{otn_list_institutions}
\alias{otn_list_institutions}
\title{Retrieve OTN institution list}
\usage{
otn_list_institutions(server = "otn")
}
\arguments{
\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}
}
\description{
Retrieve OTN institution list
}
