% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_general.R
\name{otn_list_stats}
\alias{otn_list_stats}
\title{Retrieve OTN server statistics}
\usage{
otn_list_stats(server = "otn")
}
\arguments{
\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}
}
\description{
Retrieve OTN server statistics
}
