% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{otn_login}
\alias{otn_login}
\title{Authenticate to an Ocean Tracking Network server}
\usage{
otn_login(server = NULL, set_credentials = FALSE)
}
\arguments{
\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}

\item{set_credentials}{Logical. Provide credentials for the current
session only?}
}
\description{
This function prompts you for the username and password associated with
your OTN account. This is necessary so that you may interface with any
project-specific files.
}
\details{
A pop up will appear asking for your username and password. If everything works
out, your credentials will be kept in the sessions' cookies. Your username/password
will not be saved -- this was done intentionally so that you don't accidentally
save credentials in a public script.
}
