% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_files.R
\name{otn_project_files}
\alias{otn_project_files}
\title{List OTN project files}
\usage{
otn_project_files(project, batch_size = NULL, server = NULL)
}
\arguments{
\item{project}{Character. The project code.}

\item{batch_size}{Numeric. The number of results to return. Defaults to 25.}

\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}
}
\description{
This function lists the file names, types, upload date, and URLs of OTN
project files -- basically everything you see in the \emph{Data and Metadata}
section of your project page. You will be prompted to log in if it is not a
public project.
}
