% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{otn_search_code}
\alias{otn_search_code}
\title{Search OTN database by project code.}
\usage{
otn_search_code(code, server = NULL)
}
\arguments{
\item{code}{Character. Project code for which to search. Supports partial matching.}

\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}
}
\description{
Search OTN database by project code.
}
