% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{otn_search_contact}
\alias{otn_search_contact}
\title{Search OTN database by project contact.}
\usage{
otn_search_contact(contact, server = NULL)
}
\arguments{
\item{contact}{Character. Project contact for which to search. DOES NOT support partial matching.}

\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}
}
\description{
Search OTN database by project contact.
}
