% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{otn_search_institution}
\alias{otn_search_institution}
\title{Search OTN database by institution.}
\usage{
otn_search_institution(institution, server = "otn")
}
\arguments{
\item{institution}{Character. Institution for which to search. DOES NOT
support partial matching.}

\item{server}{Character. Lowercase network code of the desired server. One of
"otn", "act", "npact", or "devel" (the OTN development server). Note that
"etn", "fact", and "glatos" are accepted, but only to produce an error and redirect you.}
}
\description{
Search OTN database by institution.
}
