% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{otn_search_institution}
\alias{otn_search_institution}
\title{Search OTN database by institution.}
\usage{
otn_search_institution(institution, server = NULL)
}
\arguments{
\item{server}{Character. URL of the OTN-style Plone server. Defaults to the
main OTN server at \url{https://members.oceantrack.org}.}

\item{contact}{Character. Institution for which to search. DOES NOT support partial matching.}
}
\description{
Search OTN database by institution.
}
