% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{otn_search_node}
\alias{otn_search_node}
\title{Search OTN database by local node.}
\usage{
otn_search_node(node, server = "otn")
}
\arguments{
\item{node}{Character. OTN node for which to search. DOES NOT support partial matching.}

\item{server}{Character. Lowercase network code of the desired server. One of
"otn", "act", "npact", or "devel" (the OTN development server). Note that
"etn", "fact", and "glatos" are accepted, but only to produce an error and redirect you.}
}
\description{
Search OTN database by local node.
}
