% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{otn_search_species}
\alias{otn_search_species}
\title{Search OTN database by project species}
\usage{
otn_search_species(species, server = "otn")
}
\arguments{
\item{species}{Character. Project species for which to search. DOES NOT
support partial matching and only recognizes scientific names of the
form "Genus species" (note capitalization).}

\item{server}{Character. Lowercase network code of the desired server. One of
"otn", "act", "npact", or "devel" (the OTN development server). Note that
"etn", "fact", and "glatos" are accepted, but only to produce an error and redirect you.}
}
\description{
Search OTN database by project species
}
