% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{otn_set_server}
\alias{otn_set_server}
\title{Set the OTN Plone instance with which you wish to interact.}
\usage{
otn_set_server(server = c("otn", "npact", "act", "fact", "devel"))
}
\arguments{
\item{server}{Character. Lowercase network code of the desired server. One of
"otn", "npact", "act", "fact", or "devel" (the OTN development server).}
}
\description{
Set the OTN Plone instance with which you wish to interact.
}
