% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otndata-package.R
\docType{package}
\name{otndata-package}
\alias{otndata}
\alias{otndata-package}
\title{otndata: Access Ocean Tracking Network Data}
\description{
An R API to access data from the Ocean Tracking Network (OTN) database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://otndata.obrien.page}
  \item \url{https://github.com/trackyverse/otndata}
  \item Report bugs at \url{https://github.com/trackyverse/otndata/issues}
}

}
\author{
\strong{Maintainer}: Michael O'Brien \email{mike@obrien.page} (\href{https://orcid.org/0000-0003-1420-6395}{ORCID})

Authors:
\itemize{
  \item Michael O'Brien \email{mike@obrien.page} (\href{https://orcid.org/0000-0003-1420-6395}{ORCID})
}

}
\keyword{internal}
